pyproject.toml
[build-system]
requires = ["hatchling>=1.5.0"] # Or a more specific version
build-backend = "hatchling.build"

[project]
name = "apu"
version = "0.1.2"
authors = [
    { name="Justin Miles", email="jumiles@paloaltonetworks.com" },
]
description = "A short description of your package."
readme = "README.md"
requires-python = ">=3.8"
classifiers = [
    "Programming Language :: Python :: 3",
    "Operating System :: OS Independent",
]
dependencies = [
    "aiohttp",
    "argparse",
    "asyncio",
    "boto3",
    "click",
    "datetime",
    "http_logging",
    "pandas",
    "requests",
    "grequests",
    "rx",
    "prismacloud-api",
    "update_checker",
    "pcpi",
    "urllib3",
    "python-dotenv",
    "pyyaml"
]

[project.urls]
"Homepage" = "https://github.com/PCS-LAB-ORG/api_python_utilities"

[tool.hatch.build]
include = [
    "*.py",
    "login.py"
]
exclude = [
    "*.pyc",
    ".ipynb_checkpoints",
    "*cred*"
]

[tool.pytest.ini_options]
testpaths = ["tests"]
addopts = [ "--junitxml=tests/results/report.xml", "--log-file=tests/results/test_logs.log", "-v"]

[tool.pylint.messages_control]
# Disable specific messages by their symbolic names or IDs.
disable = [
    "missing-module-docstring",
    "unspecified-encoding",
    "missing-timeout",
    "missing-docstring",
    "missing-module-docstring",
    "raising-bad-type",
    "logging-fstring-interpolation",
    "pointless-string-statement",
    "global-statement",
    "missing-function-docstring",
    "invalid-name", # I may be able to narrow this
    "line-too-long",
    "broad-exception-caught",
    "redefined-outer-name",
    "unnecessary-negation",
    "broad-exception-raised"
]
[tool.pylint.format]
# Set the max line length
max-line-length = 500
